地址簿操作添加审计

+++ Connector/lib/api/useraddr.pm
@@ -29,6 +29,8 @@ post '/useraddr' => sub {
     )->check( %$param );
     return  +{ stat => $JSON::false, info => "check format fail $error" } if $error;
 
+    eval{ $api::mysql->execute( "insert into openc3_connector_auditlog (`user`,`title`,`content`) values('$ssouser','CREATE USERADDR','USER:$param->{user} EMAIL:$param->{email} PHONE:$param->{phone}')" ); };
+
     eval{ $api::mysql->execute( "replace into openc3_connector_useraddr (`user`,`email`,`phone`,`edit_user`) values('$param->{user}','$param->{email}','$param->{phone}', '$ssouser')" ); };
     return $@ ? +{ stat => $JSON::false, info => $@ } : +{ stat => $JSON::true, info => 'ok' };
 };
@@ -40,6 +42,13 @@ del '/useraddr/:id' => sub {
     my $param = params();
     my $error = Format->new( id => qr/^\d+$/, 1 )->check( %$param );
     return  +{ stat => $JSON::false, info => "check format fail $error" } if $error;
+
+    my @col = qw( id user email phone edit_user edit_time );
+    my $addr = eval{ $api::mysql->query( sprintf( "select %s from `openc3_connector_useraddr` where id='$param->{id}'", join( ',', @col ) ), \@col ) };
+
+    return +{ stat => $JSON::false, info => "nofind the id" } unless $addr && @$addr > 0;
+
+    eval{ $api::mysql->execute( "insert into openc3_connector_auditlog (`user`,`title`,`content`) values('$ssouser','DELETE USERADDR','USER:$addr->[0]{user} EMAIL:$addr->[0]{email} PHONE:$addr->[0]{phone}')" ); };
 
     eval{ $api::mysql->execute( "delete from openc3_connector_useraddr where id='$param->{id}'" ); };
     return $@ ? +{ stat => $JSON::false, info => $@ } : +{ stat => $JSON::true };
